+++ src/components/Menu.tsx
@@ -1,0 +1,9 @@
+import useMenu from "../hooks/useMenu";
+import { Menu as MUIMenu, MenuItem } from "@mui/material";
+
+const Menu = () => {
+  // const { anchorEl, open, handleClick, handleClose } = useMenu();
+  return <Menu></Menu>;
+};
+
+export default Menu;