+++ src/store/taskStore.ts
@@ -16,15 +16,22 @@ type actions = {
 };
 
 const useTaskStore = create<state & actions>()((set) => ({
+  // Stores all task items; they are objects of type taskType
   tasks: [],
+
+  // Add a task
   addTask: (task: taskType) =>
     set((state) => ({
       tasks: [...state.tasks, task],
     })),
+
+  // Remove a task by id
   removeTask: (taskId: string) =>
     set((state) => ({
       tasks: state.tasks.filter((task: taskType) => task.id !== taskId),
     })),
+
+  // Toggle a task as complete/incomplete by id
   toggleComplete: (taskId: string) =>
     set((state) => ({
       tasks: state.tasks.map((task: taskType) => {
@@ -34,6 +41,8 @@ const useTaskStore = create<state & actions>()((set) => ({
         return task;
       }),
     })),
+
+  // Remove all completed tasks at once
   removeCompletedTasks: () =>
     set((state) => ({
       tasks: state.tasks.filter((task: taskType) => !task.isComplete),