+++ src/tests/Home.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import TaskForm from "../pages/TaskForm";
+import TaskForm from "../pages/TaskEntryPage";
 import userEvent from "@testing-library/user-event";
 import { createRoutesStub } from "react-router";
 import useTaskStore from "../store/taskStore";
-import Home from "../pages/Home";
+import Home from "../pages/HomePage";
 
 const originalState = useTaskStore.getState();