+++ src/tests/TaskForm.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
-import { describe, it, expect } from "vitest";
-import TaskForm from "../pages/TaskForm";
+import { describe, expect } from "vitest";
+import TaskForm from "../pages/TaskEntryPage";
 import userEvent from "@testing-library/user-event";
 import { createRoutesStub } from "react-router";
 import useTaskStore from "../store/taskStore";