+++ src/utils/fullRoutes.ts
@@ -1,0 +1,6 @@
+import { FRONT, ADD_TASK, EDIT_TASK } from "./routeSegments";
+
+// All valid routes within the application
+export const HOME_ROUTE = `/${FRONT}`;
+export const ADD_TASK_ROUTE = `/${FRONT}/${ADD_TASK}`;
+export const EDIT_TASK_ROUTE = `/${FRONT}/${EDIT_TASK}`;