+++ src/utils/paths.ts
@@ -1,4 +1,0 @@
-// export const PATH_FRONT = "/Todo-App/";
-export const PATH_FRONT = "/Todo-App";
-export const HOME = `${PATH_FRONT}`;
-export const ADD_TASK = `${PATH_FRONT}/add-task`;